fix: Remove service from unit test
To avoid running the main function of the service, which fails with process.exit and stops the unit tests

No test only service helpers.
